Add createOrUpdate mutations generators (#24)
# Description ✍️

This PR introduces a new feature in the GraphQL library, adding a
`CreateOrUpdate` mutation, which allows users to create or update
records in a single operation. It also includes minor refactoring to
improve code readability and consistency.

#### New Feature

- **CreateOrUpdate Mutation**: A new mutation has been added to the
GraphQL library, allowing users to create or update records in a single
operation. This new mutation can be used to simplify client-side code
and reduce the number of separate mutations needed for create and update
operations.

#### Changes

1. Added `create_or_update` mutation in
`lib/graphoid/mutations/create_or_update.rb`. The mutation takes two
arguments: `where` and `data`. If a record matching the `where`
conditions is found, the record will be updated with the given `data`.
Otherwise, a new record will be created with the given `data`.

2. Registered the new `CreateOrUpdate` mutation in
`lib/graphoid/mutations/structure.rb`.

3. Minor refactoring in `lib/graphoid/utils.rb`, removing an extra
space.

Please review the changes and let me know if you have any questions or
concerns. Your feedback is much appreciated!

# Overview 🔍

## Example of a create or update user mutation based on where query

In this example, the mutation will search for a user with the email
"[john@example.com](mailto:john@example.com)". If a user with that email
is found, their record will be updated with the provided data
(firstName, lastName, and age). If no matching user is found, a new user
will be created with the provided data.

```gql
mutation {
  createOrUpdateUser(
    where: { email: "john@example.com" }
    data: {
      firstName: "John"
      lastName: "Doe"
      age: 30
    }
  ) {
    id
    email
    firstName
    lastName
    age
  }
}
```

# Checks ☑️

- mutation createOrUpdateModel(where: {...}, data: {...})